refactor: view -> ui 패키지 변
