+++ android/app/src/main/java/io/coursepick/coursepick/presentation/ui/Event.kt
@@ -1,0 +1,22 @@
+package io.coursepick.coursepick.presentation.ui
+
+/**
+ * Used as a wrapper for data that is exposed via a LiveData that represents an event.
+ */
+class Event<out T>(
+    val content: T,
+) {
+    var hasBeenHandled = false
+        private set
+
+    /**
+     * Returns the content and prevents its use again.
+     */
+    fun getContentIfNotHandled(): T? =
+        if (hasBeenHandled) {
+            null
+        } else {
+            hasBeenHandled = true
+            content
+        }
+}
