+++ android/app/src/main/java/io/coursepick/coursepick/presentation/ui/SingleLiveData.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation
+package io.coursepick.coursepick.presentation.ui
 
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.MutableLiveData
@@ -28,25 +28,4 @@ abstract class SingleLiveData<T> {
     ) {
         liveData.observe(owner) { it.getContentIfNotHandled()?.let(onResult) }
     }
-}
-
-/**
- * Used as a wrapper for data that is exposed via a LiveData that represents an event.
- */
-private class Event<out T>(
-    val content: T,
-) {
-    var hasBeenHandled = false
-        private set
-
-    /**
-     * Returns the content and prevents its use again.
-     */
-    fun getContentIfNotHandled(): T? =
-        if (hasBeenHandled) {
-            null
-        } else {
-            hasBeenHandled = true
-            content
-        }
 }
